app/requirements.txt
pyodide-http==0.2.1
cognite-sdk==7.46.1
